Merging from working-draft-01 branch into main

+++ modules/CPA/cpa.nf
@@ -3,7 +3,6 @@
  * Description: This module performs consensus pathway analysis on metagenomic data.
  * Version: v0.1
  * Author: Kanishka Manna
- * Date: 2025-06-11
  */
 
 // Generate GO terms for CPA analysis
@@ -13,25 +12,36 @@ process GENERATE_GOTERMS {
     label 'go_term_conda'
     label 'medium'
 
-    publishDir "${params.output}/goterms", mode: 'copy'
+    publishDir "${params.goterm_db}", mode: 'symlink'
+
     output:
         path ("GOTerms.rds"), emit: goterms
         path ("gene2go.gz")
         path ("All_Data.gene_info.gz")
         path ("go.obo")
+        path (".done")
+    
+    when:
+        ! file("${params.goterm_db}/GOTerms.rds").exists() || 
+        ! file("${params.goterm_db}/gene2go.gz").exists() || 
+        ! file("${params.goterm_db}/All_Data.gene_info.gz").exists() || 
+        ! file("${params.goterm_db}/go.obo").exists() ||
+        ! file("${params.goterm_db}/.done").exists()
     
     script:
     """
-    # Download files if not present
-    [ -f gene2go.gz ] || wget -O gene2go.gz https://ftp.ncbi.nih.gov/gene/DATA/gene2go.gz
-    [ -f All_Data.gene_info.gz ] || wget -O All_Data.gene_info.gz https://ftp.ncbi.nlm.nih.gov/gene/DATA/GENE_INFO/All_Data.gene_info.gz
-    [ -f go.obo ] || wget -O go.obo http://purl.obolibrary.org/obo/go.obo
+    db_dir="."
+    
+    if [ ! -f "\$db_dir/.done" ]; then
+        echo "Downloading and generating GO term resources in \$db_dir"
+        [ -f gene2go.gz ] || wget -O gene2go.gz https://ftp.ncbi.nih.gov/gene/DATA/gene2go.gz
+        [ -f All_Data.gene_info.gz ] || wget -O All_Data.gene_info.gz https://ftp.ncbi.nlm.nih.gov/gene/DATA/GENE_INFO/All_Data.gene_info.gz
+        [ -f go.obo ] || wget -O go.obo http://purl.obolibrary.org/obo/go.obo
 
-    if [ ! -f GOTerms.rds ]; then
-        echo "Generating GOTerms.rds in goterms"
         goterms.R GOTerms.rds gene2go.gz All_Data.gene_info.gz go.obo
+        touch .done
     else
-        echo "GOTerms.rds already exists in goterms, skipping generation."
+        echo "GOTerms.rds already exists in \$db_dir, skipping download and generation."
     fi
     """
 }
@@ -60,13 +70,13 @@ process EXTRACT_METAINFO {
 
 
 // Process to run CPA analysis
-process CPA_ANALYSIS {
+process CONSENSUS_PATHWAY_ANALYSIS {
     tag "Consensus Pathway Analysis"
 
     label 'cpa_conda'
     label 'high'
     
-    publishDir "${params.output}/cpa_out", mode: 'copy'
+    publishDir "${params.output}", mode: 'move'
 
     input:
         path (samplesheet)
